requirements.txt
contourpy==1.0.6
customtkinter==4.6.3
cycler==0.11.0
darkdetect==0.8.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tk==0.1.0
